+++ custom_components/mydolphin_plus/component/api/mydolphin_plus_api.py
@@ -9,6 +9,7 @@ import logging
 import os
 import sys
 import uuid
+from typing import Callable
 
 from AWSIoTPythonSDK.MQTTLib import AWSIoTMQTTClient
 import aiohttp
@@ -45,7 +46,9 @@ class MyDolphinPlusAPI:
     awsiot_id: str | None
     awsiot_client: AWSIoTMQTTClient | None
 
-    def __init__(self, hass: HomeAssistant | None, config_data: ConfigData):
+    callback: Callable[[], None]
+
+    def __init__(self, hass: HomeAssistant | None, config_data: ConfigData, callback: Callable[[], None] | None = None):
         try:
             self._last_update = datetime.now()
             self.hass = hass
@@ -66,6 +69,7 @@ class MyDolphinPlusAPI:
             self.awsiot_id = None
             self.awsiot_client = None
 
+            self.callback = callback
             self.data = {}
 
         except Exception as ex:
@@ -123,7 +127,6 @@ class MyDolphinPlusAPI:
             _LOGGER.error(
                 f"Failed to validate login to MyDolphin Plus API ({self.base_url}), error: {ex}, line: {line_number}"
             )
-
 
     def _validate_request(self, endpoint):
         if not ConnectivityStatus.is_api_request_allowed(endpoint, self.status):
@@ -205,7 +208,7 @@ class MyDolphinPlusAPI:
         if self.status == ConnectivityStatus.Failed:
             await self.initialize()
 
-        self._load_mqtt_details()
+        self._refresh_details()
 
     async def _login(self):
         await self._service_login()
@@ -220,7 +223,7 @@ class MyDolphinPlusAPI:
 
         self._listen()
 
-        self._load_mqtt_details()
+        self._refresh_details()
 
     async def _service_login(self):
         try:
@@ -286,7 +289,7 @@ class MyDolphinPlusAPI:
             _LOGGER.error(f"Failed  to retrieve AWS token from service, Error: {str(ex)}, Line: {line_number}")
             self.status = ConnectivityStatus.Failed
 
-    def _load_mqtt_details(self):
+    def _refresh_details(self):
         get_topic = TOPIC_GET.replace("/#", "").replace("{}", self.serial)
 
         self.awsiot_client.publish(get_topic, None, 0)
@@ -473,7 +476,7 @@ class MyDolphinPlusAPI:
 
     def _publish(self, topic, message):
         if self.status == ConnectivityStatus.Connected:
-            self.awsiot_client.publish(topic, message, 1)
+            self.awsiot_client.publish(topic, message, 0)
 
         else:
             _LOGGER.error(f"Failed to publish message: {message} to {topic}")
@@ -485,21 +488,24 @@ class MyDolphinPlusAPI:
 
             payload = json.loads(message_payload)
 
-            _LOGGER.debug(f"Message received for device {self.serial}, Topic: {message_topic}, Payload: {payload}")
+            _LOGGER.info(f"Message received for device {self.serial}, Topic: {message_topic}")
 
-            if message_topic.endswith("/accepted"):
+            if message_topic.endswith("update/accepted"):
+                self._refresh_details()
+
+            elif message_topic.endswith("get/accepted"):
                 state = payload.get("state", {})
                 reported = state.get("reported", {})
-
-                is_connected = reported.get("isConnected", {})
-                self.data["isConnected"] = is_connected.get("connected", False)
 
                 for category in reported.keys():
                     category_data = reported.get(category)
 
                     if category_data is not None:
-                        _LOGGER.info(f"{category} - {category_data}")
+                        _LOGGER.debug(f"{category} - {category_data}")
                         self.data[category] = category_data
+
+                if self.callback is not None:
+                    self.callback()
 
         except Exception as ex:
             exc_type, exc_obj, tb = sys.exc_info()
@@ -545,13 +551,34 @@ class MyDolphinPlusAPI:
 
         return result
 
-    async def _send_desired_command(self, payload):
-        _LOGGER.warning(f"Sending command is not implemented yet")
+    async def _send_desired_command(self, payload: dict | None):
+        update_topic = TOPIC_UPDATE.replace("/#", "").replace("{}", self.serial)
+
+        new_state = {
+            "state": {
+                "desired": payload
+            }
+        }
+
+        request_data = json.dumps(new_state)
+
+        if self.status == ConnectivityStatus.Connected:
+            self.awsiot_client.publish(update_topic, request_data, 0)
+
+        else:
+            _LOGGER.error(f"Failed to publish message: {new_state} to {update_topic}")
 
     async def set_cleaning_mode(self, cleaning_mode):
-        _LOGGER.warning(f"Set cleaning mode is not implemented yet, Value: {cleaning_mode}")
+        data = {
+            "cycleInfo": {
+                "cleaningMode": {
+                    "mode": cleaning_mode
+                }
+            }
+        }
 
-        await self._send_desired_command(None)
+        _LOGGER.info(f"Set cleaning mode, Desired: {data}")
+        await self._send_desired_command(data)
 
     async def set_delay(self,
                         device: str,
@@ -594,8 +621,7 @@ class MyDolphinPlusAPI:
             }
         }
 
-        _LOGGER.warning(f"Set schedule is not implemented yet, Desired: {request_data}")
-
+        _LOGGER.info(f"Set schedule, Desired: {request_data}")
         await self._send_desired_command(request_data)
 
     async def set_led_mode(self, mode: int):
@@ -606,11 +632,14 @@ class MyDolphinPlusAPI:
         }
 
         request_data = self.data.get("led", default_data)
-        request_data["ledMode"] = mode
+        request_data["ledMode"] = int(mode)
 
-        _LOGGER.warning(f"Set led mode is not implemented yet, Desired: {request_data}")
+        data = {
+            "led": request_data
+        }
 
-        await self._send_desired_command(request_data)
+        _LOGGER.info(f"Set led mode, Desired: {data}")
+        await self._send_desired_command(data)
 
     async def set_led_intensity(self, intensity: int):
         default_data = {
@@ -622,9 +651,12 @@ class MyDolphinPlusAPI:
         request_data = self.data.get("led", default_data)
         request_data["ledIntensity"] = intensity
 
-        _LOGGER.warning(f"Set led intensity is not implemented yet, Desired: {request_data}")
+        data = {
+            "led": request_data
+        }
 
-        await self._send_desired_command(request_data)
+        _LOGGER.info(f"Set led intensity, Desired: {data}")
+        await self._send_desired_command(data)
 
     async def set_led_enabled(self, is_enabled: bool):
         default_data = {
@@ -636,9 +668,12 @@ class MyDolphinPlusAPI:
         request_data = self.data.get("led", default_data)
         request_data["ledEnable"] = is_enabled
 
-        _LOGGER.warning(f"Set led enabled mode is not implemented yet, Desired: {request_data}")
+        data = {
+            "led": request_data
+        }
 
-        await self._send_desired_command(request_data)
+        _LOGGER.info(f"Set led enabled mode, Desired: {data}")
+        await self._send_desired_command(data)
 
     async def drive(self, device: str, direction: str):
         if device != self.serial:
@@ -646,7 +681,7 @@ class MyDolphinPlusAPI:
 
         _LOGGER.warning(f"Drive is not implemented yet, Value: {direction}")
 
-        await self._send_desired_command(None)
+        # await self._send_desired_command(None)
 
     async def pickup(self, device: str):
         if device != self.serial:
@@ -654,7 +689,7 @@ class MyDolphinPlusAPI:
 
         _LOGGER.warning(f"Pickup is not implemented yet")
 
-        await self._send_desired_command(None)
+        # await self._send_desired_command(None)
 
     async def set_power_state(self, is_on: bool):
         request_data = None
@@ -666,6 +701,5 @@ class MyDolphinPlusAPI:
                 }
             }
 
-        _LOGGER.warning(f"Set power state is not implemented yet, Desired: {request_data}")
-
+        _LOGGER.info(f"Set power state, Desired: {request_data}")
         await self._send_desired_command(request_data)